MAINT: Fix test output for test_text_extraction_layout_mode
The Font class calculates default widths slightly different from
the code in the build_font_width_map method in _cmap.py. The main
difference is that the old code considers all non-zero widths in
the widths array, whereas the new code considers the subset that
actually exists in the character map.

The output contains a few more spaces, mostly returning two spaces
between words instead of one. Mostly, however, the output is the
same.

Most important difference:
-tincidunturna. Nullaullamcorpervestibulumturpis.
+tincidunt urna. Nulla ullamcorper vestibulum turpis.

This shows that the new output is better than the old one. However,
it still has this as well:
hachabitasseplateadictumst. Integertempusconva-

This shows that our approach of averaging widths to calculate
default width can still be improved.